refactor(app): AuthContext dédié, handler PayDunya isolé, EntitlementsProvider autour de la nav connectée, App.tsx simplifié; cleanup(LoginScreen)
