+++ package.json
@@ -48,6 +48,7 @@
   "devDependencies": {
     "@babel/core": "^7.25.2",
     "@types/react": "~19.1.10",
+    "@types/react-native": "^0.72.8",
     "@types/use-sync-external-store": "^1.5.0",
     "typescript": "~5.9.2"
   },
