+++ src/contexts/AuthContext.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export const AuthContext = React.createContext<{
+  user: any;
+  setUser: (u: any) => void;
+}>({
+  user: null,
+  setUser: () => {},
+});
+
+
